STUDIO-15: record the composite stream
Composite stream recording finally added. The `Compositor` class now contains
a `Recorder` object used for, well, recording the composite stream.

`App` subsequently listens for `Compositor`'s `record.prepare` and `record.complete`
events, and lists the recording results on completion as it does `DeviceManager`'s.
A `Composite` flavor has been added when downloading the `Compositor` stream.

Positioning of streams in the composite stream is done as follows:
1. If it is the desktop stream, position it across the entire canvas (full width),
   centred vertically.
2. If it is a non-desktop video stream, place it in the bottom right corner.It is
   scaled down to either a third of the canvas' width or height depending on the
   stream's aspect ratio.

The rendering of the composite stream is done in the main RequestAnimationFrame
loop; this means that if the recorder tab loses focus, composite stream recording,
for all intents and purposes, comes to a halt. A potential work around would be
to use a WebAudio API timed loop and/or WebGL contexts for recording, more [here](https://stackoverflow.com/questions/44156528/canvas-doesnt-repaint-when-tab-inactive-backgrounded-for-recording-webgl).

`RAFLoop` now has the option to pause/resume execution of subscriptions. This is
useful as the audio analyser adapts to whether the user views the merged or non-merged stream.