Move "impressum" content to "about" page
The "legal notices" link at the very bottom was annoying and made the
page look worse. It is also nothing the normal user is interested in,
so "hiding" that information in the "about" page is a good idea.

The `Footer` was not used for anything else. And since I don't think we
will need a footer anytime soon, I removed it as well.

Fixes #200