Fix bug when using `useRef`
The `current` value is initialized to the argument given to `useRef`.
Since nothing is given to that function, `current` is `undefined`.
Before accessing the `srcObject` property, we now just check if the
`current` field has been initialized already.

This change partially fixes the app in Edge: now at least the UI is
shown. There are still other bugs and one can't yet see the webcam
preview video.