Fix bug when using `useRef` to set `video.srcObject`
The `current` value is initialized to the argument given to `useRef`.
Since nothing is given to that function, `current` is `undefined`.
Before accessing the `srcObject` property, we now just check if the
`current` field has been initialized already.

Additionally, when assigning to `srcObject`, Edge makes sure the value
is not `undefined` (which is the case until the user actually approved
capturing their camera). So on Edge this broke, while the other
browsers just ignored that. With this change, Edge "works" in the
beginning: it looks normal and you can even preview your camera image.
However, since `MediaRecorder` is not implemented, the user can't
record.