Add "supported browsers" table to README
This table specifies which features are working for each browser/system
combination we might want to support currently. This is more
informative than the old "officially supported" list, which was super
inaccurate. Yes, we have to update this table from time to time, but
I don't expect that to be a lot of work.