Updates Languages/Language Selector
This patch drops the flags since they make it harder to add remove
languages. People should not select/change their language very often and
it should be sufficient to have the language name for them to make their
choice.

Even more important, the usage of the current set of flags is in
violation of their license. We would need to add attribution links if we
wanted to keep them or replace them with different flags. Another good
reason for dropping them ;-)

This also adds a few additional languages with good translation coverage
on Crowdin.