Add language detection and correct loading from local storage
Before this commit the language stored in the settings object was not
correctly used. Only when actually changing the language to another
one, the view was updated. This commit fixes that: the language the
user chose is stored in local storage and correctly retrieved. The app
is immediately displayed in the correct language.

This is done via 'i18next-browser-languagedetector'. The `language` key
is removed from our own `ocStudioSettings` object as it's not used
there anymore.

'i18next-browser-languagedetector' also provides language detection
by checking `navigator.language`. This should set the correct language
for almost all users from the beginning.