Fix "black video preview" bug in mobile Firefox
I'm not sure about the exact cause of this problem, but it has
something to do with initially loading the data. By setting the
`currentTime` to any value after sufficient data has been loaded, the
video preview works. The `onPlayHandler` handler fires almost
immediately as far as I can see.