Remove "first visit" settings
We decided that the app should work without a working OC connection.
Either because the user wants to simply download the video or because
the OC connection can only be specified later. Shortly, the user will
also be able to configure the OC connection after recording something.

The language is guessed from the user's browser, so should be correct
almost all of the time. The fallback is English so we expect almost all
users to find the settings to change the language.

With other changes to the settings page, having another "first visit"
settings page is extra code not worth a lot.