Do not exit settings page after saving, instead show success icon
Previously, saving the connection settings would forward the user back
to the previous page. That's not really what we want. Now the testing
and saving of the connection settings are indicated via icons and the
user stays on the settings page. Way more predictable behavior. The
"back" link is also removed, as most webpages do not offer it. The user
is free to leave at any time by using the navigation menu.