Let the browser decide the best MIME-type
For one, currently we prefer vp9 over other codecs. Chrome on Android
does support vp9, but encoding is very slow. This leads to ~2fps
recordings on non-powerful devices, which is obviously unacceptable.
Second, Safari only supports a few special MIME-types. We could special
case something for Safari, but we can also let the browser choose the
MIME-type. This is a good default.

We might want to add "expert settings" later with which advanced users
can tweak these things.