Improve div-structure of whole page
Before, the div structure was a bit of a mess which made it
particularly difficult to work with flexboxes. This commit used the
following structure for all pages:

- <div>: flexbox
  - <header>  (no changes here)
  - <main>: flexbox
    - <div>: Notifications
      - <div>...: one per notification
    - <div>: page contents, defined by the component for that route

With that, components can decide whether they want to simply overflow
the page (like with the "about" page) or to restrict themselves to
the screen height.