Try really hard to restrict page size to 100vh for source selection
This was way more tricky than expected. The difficult part was to find
a solution where the "next" button would be just below the preview
video instead of at the bottom of the page, even on very large screens.
We couldn't go with the "flex-grow" based solution used in the
recording screen. Instead, a "flex-shrink" based solution is used.

Right now, an outer body div has `min-height: 350px`. So sometimes
users might have to scroll. But this should only happen on very small
screens (or phones in landscape mode).