Request webcam before display share
If sharing a specific window from the desktop, some browsers bring
that window into the foreground. If it does not overlap the browser
window, users might be confused as the webcam is not requested. This
only happens after the browser window is focused again. Changing the
order works around this problem.