Improve initial loading indicator
With this commit, the loading indicator is embedded in the static HTML
meaning it is visible after the first HTTP request returned =>
basically instantaneously. This is an improvement for fast connections,
too. For one, there is never a completely white page, reducing the
number of "page changes" from two (blank -> spinner, spinner -> app) to
one (spinner -> app). Also, the JS just needs some time to load, which
might even be longer than the actual network loading times.

This commit makes the code a bit simpler, which is also a plus.