Fix error when recording data has size 0
This probably only happens if the device is really weak and the
recording is very short. In that case, no actual recording data has
been created yet. Before this commit, this would just lead to a JS
error. With this commit, no JS error occurs, a big error is shown and
the user can proceed normally (which doesn't make a lot of sense tho).