Add aspect ratio selector for the camera stream
Having tested this with a number of browser and devices, it looks like
this doesn't work most of the time. Right now we request it with
`ideal`, but even with `exact` it doesn't seem to make a difference. So
I assume we will disable that again as otherwise users are too easily
confused. However, most code in this commit is also used for the
quality selector.