Add explanation about non-supported preferences
I asked a few other people about those texts, including a certainly
technically non-versed user :P We tweaked the texts and with this,
everything should be pretty clear to the user. Particularly the "if in
doubt" part of the text really helps as inexperienced users actually
stick to those advices quite well.