Add TOML parser for settings and change default settings file to `settings.toml`
This is a breaking change! The configuration file can still be
overwritten by `REACT_APP_SETTINGS_PATH` which means that if this
variable is defined, this change is backwards compatible. This is
important for Opencast 8.x.

JSON is not a good format for a configuration file. In particular, we
want to add some long strings to the configuration at some point, which
is particularly ugly in JSON.

Viable alternatives are YAML and TOML. While YAML is more familiar to
many people in the Opencast community, TOML was chosen for its
simplicity. This helps humans understand the config file at a glance
and allows for very small parsers. The TOML parser used here is around
8kB minified+gzipped while the alternative YAML parser would be around
40kB minified+gzipped. I doubt that anyone in the Opencast community
would have problems understanding or changing TOML files.