BREAKING: pass full `user` and `lti` object to ACL template
This removes the `userName`, `userRole` and `ltiCourseId` variables
from the template view. This is a breaking change.

Before, we prepared special variables to pass as view to the template.
By passing all the information we have (the user and lti object), we
do not need to prepare a special variable for every use case. This way,
people configuring the template have more freedom. We still prepare
some convenience variables.

This is possible by modifying the Mustache.js escape function.

Two potential disadvantages:
- "People configuring the template can break stuff": they already could
  do that before. Configuring this template requires knowledge of
  Opencast and Studio; that will not change.
- "People configuring the template could get access to sensible
  information": (a) neither `/info/me.json` nor `/lti` should contain
  any sensible information (such as passwords). Furthermore, the users
  of a particular Studio installation need to trust that installation
  anyway: it could serve bad JS code anyway.

So I think this simplifies stuff for Studio.