Make metadata fields in upload form configurable
This is useful when using a custom DCC. Maybe some users want to add
a fixed value there, meaning that the form field needs to be hidden. Or
users want to make one field optional and fill it with some other value
in the case the user didn't specify a value.

Default behavior doesn't change.