BREAKING: Remove `default*Roles` from ACL template view
The name "default" seems to imply those roles are used by default,
which is not the case. We internally discussed that making those roles
the default roles in the ACL is also not a good idea. Additionally,
these particular roles are not even that useful as a set. In most cases
a slightly different set is needed.

So we just remove them. They were just convenient variables in the view
anyway and you can still manually achieve the same effect.