Fix too long filenames
If a title is extremely long (can easily happen if it is generated
based on some metadata) Studio may generate a filename which exceeds the
file systems maximum file length. This can cause problems when handling
or processing the file.

This patch sets an upper limit of 50 characters for title and presenter
each in the filename. This should suffice to identify the name while
avoiding potential problems.