Update to react-scripts 4
This required two fixes/workarounds:

For one, our `.eslintrc` was not working anymore, complaining about how
it cannot find "react-app" to extend. That's because that config,
`eslint-config-react-app` was not stored directly in `node_modules`
but inside of `node_modules/react-scripts/node_modules`. Thus, eslint
could not find it anymore. The solution is to have it as direct
dependency.

The next problem is due to Reacts new runtime. `theme-ui` only works
with the classic runtime right now, so for now the new runtime is
simply disabled.