Fix master deployment by not removing build artifacts
The problem is, that we want the test deployment and the
production deployment for `master`. But both workflows delete
the build artifacts after downloading them. So only one gets it,
the other one fails.
This is solved by just not deleting build artifacts.

The build artifacts for Studio are very small (1.4MB) or so, so
even with lots of parallel builds, it's unlikely we will reach
GitHub's limit.