Fix prototype pollution vulnerability in settings handling
URL parameters could be used to modify the root prototype, eg.:
   ?__proto__.toString=peter

Fortunately, the assigned value is always a string, so it's not possible
to overwrite functions with other functions. The example above just
leads to a JS error which is caught by the application and an error is
then displayed. It might be possible to add a new field to the prototype
that then leads to other code checking `if ('foo' in obj)` to behave
differently. But I couldn't find a spot that could be abused. I might
very well have missed something, but my guess would be that this
vulnerability is not really exploitable in a bad way.

Note: I think only one of the `{}` needs to be replaced by
`Object.create(null)`, but it doesn't hurt just replacing all of them.
From a quick glance, I also didn't find any similar problems in the
code base.