Address css/styling related review comments
At a screen width of 742px only the icons are now displayed in the header menu and the text is hidden.
The 742px are based on the French language (the header is the widest in French) and at a screen width of 742px the last header entry ("Info") disappears.

A group of shortcuts no longer has a fixed width.
EXCEPT from a screen width of 350px. At about this width, it starts to look too tight and squashed. Therefore the groups are then 300px wide.
So it still looks nice with a width of 320px (300px for shortcuts and 10px margin per side).