Switch from single-quote in JSX to double-quotes
This is closer to normal HTML and is also the style used in other of
our projects, e.g. Tobira and the editor.

I would have loved to amend the commit that introduced the single-quotes
rule, but that has already been on master for some time. So we have to
live with the somewhat noisy git history here.