Add Typescript, fix imports and change file endings from .js to .tsx
We use Typescript in almost all of our other projects and think it
offers great benefits. Thus we want to use it in Opencast Studio too.
Renaming the files to use .jsx or .tsx (so that editors correctly
highlight by default) was a plan anyway.

There are still tons of type errors in the code. Those are fixed in
the next commit(s).