Remove faulty configuration about shortcuts in input/textarea/...
The `ignoreTags` config is ignored when setting `ignoreEventsCondition`.
Further, `ignoreEventsCondition` should return a boolean, which leads
to a type error and I assume the returned `undefined` was always
interpreted as `false`. But lastly, we likely just want the default
behavior of the library. The current behavior already leads to a bug:
typing 'd' inside the form fields at the end downloads the video.