Consistent Settings Style
This patch updates the settings style to make the different settings
options use the same style. In particular, this means that:

- All input fields have labels
- Labels use the same font and are no longer sometimes bold
- Input element border color and radius are now identical

The last change will also effect some other views but a homogeneous
style makes sense for the whole user interface anyway. It is just not so
obvious on other pages.

This patch also puts the settings blocks in a flex-box so that we don't
end up with extremely wide dropdowns with text only on the very left. In
most cases you don't even need to scroll any longer.