Fix theme bug where always light theme colors were used
This is... annoying. So my previous hacky solution with `colorz` was
broken. That object was always exactly what I put in, so all code
referring to it always used the light mode colors. The `colors` field
is instead replaced by theme-ui, at some point, magically, to contain
strings with `var()` references. We have to use those to get the correct
colors all the time. I tried a `get colorz() { return this.colors }`
but that didn't work either as the getter was evaluated very early, and
also only once, at which point `colors` was not transformed yet.

So I decided to just bite the bullet and search&replace all cases, using
`colors` instead. This adds some ugly `?.` accesses and loses type
safety. It's basically a `record<string, any>` here. So we won't catch
color name typos, for example.

It seems that theme-ui is just not really designed for Typescript.
Usually, you also specify colors just as strings, i.e. `bg: "primary"`
which is also not type safe.

Maybe I can improve this in the upcoming redesign work, but for now we
should just fix the bug.