Add the ability to autofill the presenter field in the upload step
This patch adds a config option to specify a preference order
of potential sources to autofill the presenter field from.
Leaving it empty is equivalent to the current behavior.
The only valid value right now is `"opencast"`,
which gets the value from `/info/me.json`.
The way the feature is built, it should be easy to extend
in the future, though.
If there is a name stored in `localStorage` already,
this currently takes precedence over (any of) the suggested value(s).

Co-authored-by: naweidner <weidner@elan-ev.de>