Remove incorrect constant duration in DC catalog
This was here since the initial commit 👌 For one, having a fixed value
doesn't make sense. But also, Opencast apparently doesn't care about
`dcterms:extent`, but uses `dcterms:temporal xsi:type="dcterms:Period"`
for durations. And it's not trivial for Studio to get the correct
location anyway, so let's just throw it out.