Allow server settings path to be set via `settingsPath` query parameter
Fixes #1014

This allows institutions to easily configure Studio for multiple use
cases, without specifying super long config values in query parameters
directly.

I am still a bit scared that this might introduce a security
vulnerability, but I tried things and thought about it and I think: it
does not. Most importantly, we make sure that the passed string is only
a single file that lives in the same directory as the 'settings.toml'
that would have been loaded anyway. This also prevents Studio from
loading config files from external sources. Further consider that all
values in the config file can be overwritten via specific query
parameters anyway! Well, except `return.allowedDomains`. So this is the
only one we care about. But given that an attacker can only chose
between all settings files that live in the same directory
as 'settings.toml'. So this should be fine, I think. Also note that we
cannot just let the query parameter specify arbitrary paths as users
can upload videos and attachments to Opencast, so they could inject a
settings file that was that is reachable via the same domain.