Eject from CRA
With CRA we are forced to annotate the JSX thing in each file. It also
has lots of other problems, from slow updates, over not being able to
configure certain things, to unexpected magic behavior.

This commit just ran `npm run eject`. This left a mess which will be
cleaned up in the next commits.