Remove `theme.customCSS` setting
In the future, we might re-introduce a similar setting that would allow
modifying the accent color. Doing that in a way that prevents people
from accidentally introducing accessibility problems is not trivial.
That's why it's not included in the main PR.
Also, this custom CSS thing was quite hacky and fragile. We never
officially supported/encouraged overriding CSS variables. So replacing
this with a less brittle version is a good idea anyway.