Configure dependabot: remove outdated entries and group simple updates
The ignored dependencies are either not used anymore or used in the
newest version, so all those ignore entries were removed. Further, we
will now use the grouped updates to decrease the number of pull requests
(and thus also test deployments). It's very safe to just group minor and
patch updates. In the rare case something fails, we can still do it
manually.