Change dark high contrast mode accent color
The previous one was a very very bright green tone. I changed it to a
very bright version of our normal accent color. It would be weird to
suddenly have green where otherwise Studio only uses blue. Now I
understand that with green you can get brighter while maintaining
high saturation, better than with blue. But I think this blue is
bright enough. Further, the only reason I noticed this is that I,
as someone with strong Protonopia did not even see that that was green
at all. I assume the accent color was just white. I think by making it
a tiny bit less bright and using blue, the chance of that happening to
other people is a lot less.