Remove browser versions in support table
The shown versions are certainly outdated and Studio does not support
those anymore. Testing which versions are supported exactly would be
a lot of work. And it's not like Studio uses the latest technology
anymore, so the exact version is not too important. Especially since
browsers do update themselves very regularly nowadays. Luckily.