Add prototypal support for camera background blurring
The purpose here is to evaluate the performance
of doing this client-side, both in terms of runtime behavior,
and quality of the resulting video.
Note that it just unconditionally blurs the background,
i.e. there is no UI to control the blur, yet.

It uses the `@shiguredo/virtual-background` package,
and while it works we probably don't want to use it "in prod,"
for several reasons:

* Its documentation is almost entirely in Japanese,
* they have some questionable rules regarding contributions,
* and it doesn't even work in every Browser.
  Specifically Firefox doesn't work at all, sadly.

There are some other libraries, but weirdly they all have
similar negative aspects. Yes, there even is another Japanese one!
So if we want this, there is likely going to be
some "manual tinkering" involved.