Fix seeking in review & trim step
Aims at fixing or at least helping with #517.

As suggested in #517, this introduces the
webm-duration-fix package to Studio. This should
fix the "Preview & Trim" step on Chrome (desktop),
allowing for playback and seeking in the recorded
video. Furthermore, the duration should now
be stored correctly in the video metadata.

This should not change behaviour in Firefox, as
it was already working in Firefox. I have not tested
other browsers.

Note that recordings downloaded directly from
Chrome may still not be playable by all and every
media player. This is not something this PR can fix.