Add ability to specify `upload.acl` concisely via object
Before, one had to always specify an XML blob, which is suuuper verbose.
With this, Studio also allows a more concise representation that's also
suitable to be included in the URL. While each string in the object is
also treated as a mustache template, with the object notation, it is not
possible to conditionally add entries. At least not yet.