Fix bug causing errors being shown when uploading
One error is shown after a successful upload (see #1231), the other
appears when clicking "submit" in certain situations. Both are caused
by an incorrect `match` migration during the appkit update.

In one `match` use, each branch only calls a function for its side
effects, returning `void`, causing the `??` afterwards to always
execute. That's different from the previous behavior where that code was
only executed when none of the arms were taken. I just added the last
possible value as arm.

In the other case, `null` was a possible value that was returned, so it
was impossible to distinguish between `logged_in` and some unknown arm.
I rewrote that code to deal with `logged_in` separately, which also
makes more sense IMO.