Remove "duration calculation" code from cut step
It was added at a time where browsers would have problems showing the
video otherwise. Users could also not seek nor see the duration. This
was always a dirty trick. I stumbled upon this as this trick causes
problems in Firefox now. Removing it solves it. And to my surprise,
it seems like the original problem is also fixed in all browsers, so
we don't need this anymore.