Remove "backup synchronization" from cut step
I added this a long time ago, when first coding this step, since I was
not sure if there were effects that could make the videos diverge. So
having some regular backup that would fix the time of one video seemed
useful. What I noticed now however, is that since (at least) the
redesign, this backup sync wasn't doing anything. That's because it
relied on `lastOrigin` to be set to the last video that received input
somehow. But this was never set! So the sync code noticed a large diff,
but would not change the time.

Given that this has been unnoticed for at least two years, the question
arises: is backup sync even necessary? As I said, I added it as a
precaution without knowing what would make it necessary. The only way I
could test this is by manually enabling the video controls in the
browser, and seeking in one video. Users could do that of course, but
I would think it's rather rare. And the result is not catastrophic in
nature, just some weird play behavior.

So this commit removes the backup sync and also the `lastOrigin`
variable, since it was never initialized anyway.